Remove obsolete and redundant project code
