+++ ui/index.html
@@ -1910,7 +1910,12 @@
 <script src="/js/chat/chat-main.js?v=chat-split-1"></script>
 <script src="/js/chat-window/chat-window-protocol.js"></script>
 <script src="/js/chat-window/chat-window-host-adapter.js"></script>
-<script src="/js/chat-window.js?v=chat-window-1"></script>
+<script src="/js/chat-window/chat-window-bootstrap.js"></script>
+<script src="/js/chat-window/chat-window-coordinator.js"></script>
+<script src="/js/chat-window/chat-window-flag-core-bridge.js"></script>
+<script src="/js/chat-window/chat-window-host-view.js"></script>
+<script src="/js/chat-window/chat-window-detached-view.js"></script>
+<script src="/js/chat-window/chat-window-main.js"></script>
 <script src="/js/benchmark-ui.js?v=revamp-2"></script>
 <script src="/js/inference-stats.js?v=1"></script>
 <script src="/js/monitor/monitor-internal.js?v=1"></script>
