+++ ui/js/chat-window/chat-window-coordinator.js
@@ -1,49 +1,17 @@
-// Chat pop-out window: verified bootstrap and ownership/recovery coordination.
-// Protocol and host-adapter contributors load first; the public facade stays here.
+// Exclusive Chat workspace ownership, transfer, and recovery. All mutable state stays per coordinator.
 (function () {
     "use strict";
-
     const root = typeof window !== "undefined" ? window : globalThis;
     const I = root.LlamaGui._chatWindowInternal;
-    const {
-        PROTOCOL, PROTOCOL_VERSION, SNAPSHOT_VERSION, SNAPSHOT_KIND, RECOVERY_VERSION,
-        DEFAULT_LOCK_NAME, DEFAULT_RECOVERY_KEY, CHAT_READONLY_SETTING_FIELDS,
-        getWindow, debug, warn, isObject, isJsonValue, cloneJson, result,
-    } = I.protocol;
-    const { deriveSettingFields, createHostAdapter } = I.hostAdapter;
+    const { PROTOCOL, PROTOCOL_VERSION, SNAPSHOT_VERSION, SNAPSHOT_KIND, RECOVERY_VERSION, DEFAULT_LOCK_NAME, DEFAULT_RECOVERY_KEY, getWindow, debug, warn, isObject, isJsonValue, cloneJson, result } = I.protocol;
+    const { deriveSettingFields } = I.hostAdapter;
+    const { getOrigin, randomId, getStorage } = I.bootstrap;
 
     function defer() {
         let resolve;
         let reject;
         const promise = new Promise((res, rej) => { resolve = res; reject = rej; });
         return { promise, resolve, reject };
-    }
-
-    function getOrigin(options) {
-        const target = getWindow(options);
-        if (options && typeof options.origin === "string") return options.origin;
-        if (target && target.location && typeof target.location.origin === "string") return target.location.origin;
-        return "";
-    }
-
-    function randomId(prefix, options) {
-        const target = getWindow(options);
-        const cryptoObject = target && target.crypto;
-        if (cryptoObject && typeof cryptoObject.randomUUID === "function") {
-            try { return `${prefix}-${cryptoObject.randomUUID()}`; } catch (error) { debug(options, "randomUUID unavailable", error); }
-        }
-        const random = Math.random().toString(36).slice(2);
-        return `${prefix}-${Date.now().toString(36)}-${random}`;
-    }
-
-    function getStorage(options) {
-        if (options && Object.prototype.hasOwnProperty.call(options, "storage")) return options.storage;
-        const target = getWindow(options);
-        if (!target) return null;
-        try { return target.localStorage; } catch (error) {
-            debug(options, "localStorage is unavailable", error);
-            return null;
-        }
     }
 
     function createCoordinator(options) {
@@ -1097,889 +1065,5 @@
         return Object.freeze(coordinator);
     }
 
-    const DETACHED_QUERY_PARAM = "chat-window";
-    const POPUP_NAME = "llama-gui-chat";
-    const POPUP_FEATURES = "popup=yes,width=960,height=760,resizable=yes,scrollbars=yes";
-    const DEFAULT_POPOUT_TITLE = "Open Chat in a separate window";
-
-    function isDetachedView(target) {
-        const current = target || (typeof window !== "undefined" ? window : null);
-        try {
-            const href = current && current.location && typeof current.location.href === "string"
-                ? current.location.href : "";
-            if (!href) return false;
-            if (typeof URL === "function") return new URL(href).searchParams.get(DETACHED_QUERY_PARAM) === "1";
-            const query = href.split("?", 2)[1]?.split("#", 1)[0] || "";
-            return query.split("&").some(part => {
-                const [key, value] = part.split("=", 2);
-                return decodeURIComponent(key || "") === DETACHED_QUERY_PARAM && decodeURIComponent(value || "") === "1";
-            });
-        } catch (error) {
-            if (current && current.console && typeof current.console.debug === "function") {
-                current.console.debug("Unable to determine Chat window mode", error);
-            }
-            return false;
-        }
-    }
-
-    function isClosedWindow(value) {
-        try { return !value || value.closed === true; } catch (error) { return true; }
-    }
-
-    function safeRead(getter, fallback, logger) {
-        try { return typeof getter === "function" ? getter() : fallback; }
-        catch (error) {
-            if (logger && typeof logger.debug === "function") logger.debug("Chat window host read failed", error);
-            return fallback;
-        }
-    }
-
-    function createFlagCoreBridge(host, options = {}) {
-        const listeners = new Set();
-        let values = {};
-        let selectedModel = "";
-        const logger = options.window && options.window.console;
-
-        function refresh(nextSettings) {
-            const settings = nextSettings && typeof nextSettings === "object" ? nextSettings
-                : safeRead(host && host.getSettings, {}, logger);
-            values = Object.assign({}, settings || {});
-            selectedModel = values.selected_model === null || values.selected_model === undefined
-                ? "" : String(values.selected_model);
-            for (const listener of Array.from(listeners)) {
-                try { listener(values); } catch (error) {
-                    if (logger && typeof logger.warn === "function") logger.warn("Detached Chat settings listener failed", error);
-                }
-            }
-            return values;
-        }
-
-        function currentValues() {
-            const live = safeRead(host && host.getSettings, null, logger);
-            if (live && typeof live === "object" && !Array.isArray(live)) refresh(live);
-            return Object.assign({}, values);
-        }
-
-        function write(patch) {
-            if (!host || typeof host.setSettings !== "function") throw new Error("Chat host settings writer is unavailable.");
-            try {
-                if (typeof host.isSessionValid === "function" && host.isSessionValid() !== true) {
-                    throw new Error("Chat host is not connected.");
-                }
-            } catch (error) {
-                options.onUnavailable?.(error);
-                return currentValues();
-            }
-            let resultValue;
-            try { resultValue = host.setSettings(patch); } catch (error) {
-                logger?.warn?.("Chat host settings write failed", error);
-                options.onUnavailable?.(error);
-                return currentValues();
-            }
-            refresh(resultValue);
-            return values;
-        }
-
-        refresh();
-        return Object.freeze({
-            getFlagValues: currentValues,
-            getSelectedModel: () => {
-                currentValues();
-                return selectedModel;
-            },
-            getCurrentTool: () => "llama-server",
-            setFlagValue: (field, value) => write({ [field]: value }),
-            setMultipleFlagValues: patch => write(patch),
-            subscribe(listener) {
-                if (typeof listener !== "function") return () => {};
-                listeners.add(listener);
-                return () => listeners.delete(listener);
-            },
-            refresh,
-        });
-    }
-
-    function waitForPeer(coordinator, timeoutMs = 10000) {
-        if (coordinator.isPeerVerified()) return Promise.resolve(true);
-        const limit = Number.isFinite(Number(timeoutMs)) ? Math.max(0, Number(timeoutMs)) : 10000;
-        return new Promise(resolve => {
-            let finished = false;
-            const finish = value => {
-                if (finished) return;
-                finished = true;
-                if (unsubscribe) unsubscribe();
-                if (timer) clearInterval(timer);
-                resolve(Boolean(value));
-            };
-            const unsubscribe = coordinator.subscribe(() => {
-                if (coordinator.isPeerVerified()) finish(true);
-            });
-            const timer = typeof setInterval === "function" ? setInterval(() => {
-                if (coordinator.isPeerVerified()) finish(true);
-            }, 50) : null;
-            if (typeof setTimeout === "function") setTimeout(() => finish(coordinator.isPeerVerified()), limit);
-            else finish(coordinator.isPeerVerified());
-        });
-    }
-
-    function whenDomReady(target) {
-        if (!target || !target.document || target.document.readyState !== "loading") return Promise.resolve();
-        return new Promise(resolve => target.document.addEventListener("DOMContentLoaded", resolve, { once: true }));
-    }
-
-    function showDetachedError(target, title, message, options = {}) {
-        const doc = target && target.document;
-        const placeholder = doc?.getElementById("chat-window-placeholder");
-        const layout = doc?.getElementById("chat-layout");
-        const heading = placeholder?.querySelector("h3");
-        const detail = placeholder?.querySelector("p");
-        const showWindow = doc?.getElementById("btn-chat-show-window");
-        const returnHere = doc?.getElementById("btn-chat-return-here");
-        const fullGui = doc?.getElementById("chat-window-open-full-gui");
-        const headerActions = doc?.querySelectorAll("#section-chat .chat-header-actions button");
-        if (heading) {
-            heading.textContent = title;
-            heading.tabIndex = -1;
-        }
-        if (detail) detail.textContent = message;
-        if (layout) layout.hidden = true;
-        if (placeholder) placeholder.hidden = false;
-        if (showWindow) {
-            showWindow.hidden = true;
-            showWindow.disabled = true;
-            showWindow.setAttribute("aria-disabled", "true");
-        }
-        if (returnHere) {
-            returnHere.hidden = true;
-            returnHere.disabled = true;
-            returnHere.setAttribute("aria-disabled", "true");
-        }
-        headerActions?.forEach(button => {
-            button.disabled = true;
-            button.setAttribute("aria-disabled", "true");
-        });
-        if (fullGui) {
-            fullGui.hidden = options.fullGuiLink !== true;
-            if (options.fullGuiLink === true) {
-                try {
-                    const url = new URL(target.location.href);
-                    url.searchParams.delete(DETACHED_QUERY_PARAM);
-                    fullGui.href = url.href;
-                } catch (error) {
-                    fullGui.hidden = true;
-                    target.console?.debug?.("Unable to build full GUI recovery link", error);
-                }
-            }
-        }
-        doc?.body?.setAttribute("data-chat-window-error", "true");
-        heading?.focus?.();
-    }
-
-    function startHostView(options = {}) {
-        const target = options.window || (typeof window !== "undefined" ? window : null);
-        const chatUi = options.chatUi || target?.LlamaGui?.chatUi;
-        const flagCore = options.flagCore || target?.LlamaGui?.flagCore;
-        if (!target || !chatUi || !flagCore) return Promise.resolve(result(false, "chat-host-unavailable"));
-        if (api._hostView) return api._hostView.ready;
-
-        let popup = null;
-        let popupProof = null;
-        let detached = false;
-        let mainLayout = null;
-        let originalFocus = null;
-        let closedCheckTimer = null;
-        const storage = getStorage({ window: target });
-        const logger = target.console;
-        const getStatus = () => safeRead(options.getLatestStatus, null, logger);
-        const getLifecycle = () => safeRead(options.getLifecycleSnapshot, null, logger);
-        function setHostStatus(message) {
-            const status = target.document?.getElementById("chat-window-host-status");
-            if (!status) return;
-            status.textContent = message || "";
-            status.hidden = !message;
-        }
-        const hostAdapter = createHostAdapter({
-            window: target,
-            storage,
-            flagCore,
-            getChatSamplerFlagIds: () => chatUi.getChatSamplerFlagIds?.() || [],
-            getSelectedModel: () => typeof flagCore.getSelectedModel === "function"
-                ? flagCore.getSelectedModel() : undefined,
-            getActiveRuntime: getLifecycle,
-            getStatus,
-            getInference: () => safeRead(options.getInferenceSnapshot, null, logger),
-            resetInferenceBaseline: options.resetInferenceBaseline,
-            getAuthorizationHeaders: options.getApiAuthorizationHeaders,
-            bringToFront: () => {
-                try { target.focus(); return true; } catch (error) {
-                    logger?.debug?.("Main Chat focus failed", error);
-                    return false;
-                }
-            },
-            navigate: tabId => {
-                if (typeof options.switchTab !== "function") return false;
-                try {
-                    options.switchTab(tabId);
-                    try { target.focus(); } catch (error) { logger?.debug?.("Main Chat focus failed", error); }
-                    return true;
-                } catch (error) { return false; }
-            },
-        });
-        const coordinator = createCoordinator({
-            window: target,
-            chatUi,
-            flagCore,
-            hostAdapter,
-            verifyPeerProof: payload => Boolean(popupProof && payload && payload.proof
-                && payload.proof.key === popupProof.key && payload.proof.value === popupProof.value),
-        });
-
-        function notifyHostChange(change) {
-            if (change && change.type === "session-invalidated") {
-                return hostAdapter.invalidateSession();
-            }
-            return hostAdapter.notify(change);
-        }
-
-        function setButtonDisabled(button, disabled) {
-            if (!button) return;
-            const value = Boolean(disabled);
-            button.disabled = value;
-            button.setAttribute("aria-disabled", String(value));
-        }
-
-        function canFocus(element) {
-            if (!element || element.hidden || element.disabled || element.getAttribute?.("aria-disabled") === "true"
-                || element.isConnected === false || typeof element.focus !== "function") return false;
-            if (typeof element.getClientRects === "function") {
-                try {
-                    const rects = element.getClientRects();
-                    if (rects && rects.length === 0) return false;
-                } catch (error) {
-                    logger?.debug?.("Unable to inspect Chat focus target visibility", error);
-                    return false;
-                }
-            }
-            return true;
-        }
-
-        function focusFirst(...elements) {
-            for (const element of elements) {
-                if (!canFocus(element)) continue;
-                try {
-                    element.focus();
-                    if (!target.document || !target.document.activeElement || target.document.activeElement === element) return true;
-                } catch (error) {
-                    logger?.debug?.("Chat focus target failed", error);
-                }
-            }
-            return false;
-        }
-
-        function focusPlaceholderAction() {
-            return focusFirst(
-                target.document.getElementById("btn-chat-show-window"),
-                target.document.getElementById("btn-chat-return-here"),
-            );
-        }
-
-        function focusMainChatFallback() {
-            const body = target.document?.body;
-            const documentElement = target.document?.documentElement;
-            const previousFocus = originalFocus && originalFocus !== body && originalFocus !== documentElement
-                ? originalFocus : null;
-            return focusFirst(
-                target.document.getElementById("chat-input"),
-                previousFocus,
-                target.document.getElementById("btn-chat-new"),
-                target.document.getElementById("btn-open-history"),
-                target.document.getElementById("btn-chat-popout"),
-            );
-        }
-
-        function isLiveDetached() {
-            return detached && !isClosedWindow(popup);
-        }
-
-        function notifyDetachedChange() {
-            if (typeof options.onDetachedChange !== "function") return;
-            try { options.onDetachedChange(isLiveDetached()); } catch (error) { logger?.warn?.("Chat detached-state callback failed", error); }
-        }
-
-        function setDetachedUi(active, reason) {
-            detached = Boolean(active);
-            const state = coordinator.getState();
-            const placeholder = target.document.getElementById("chat-window-placeholder");
-            const layout = target.document.getElementById("chat-layout");
-            const popout = target.document.getElementById("btn-chat-popout");
-            const returnHere = target.document.getElementById("btn-chat-return-here");
-            const showWindow = target.document.getElementById("btn-chat-show-window");
-            if (placeholder) placeholder.hidden = !detached;
-            if (layout) layout.hidden = detached;
-            if (popout) {
-                setButtonDisabled(popout, detached || !coordinator.isOwner() || !state.popoutAvailable);
-                if (reason) popout.title = reason;
-            }
-            if (returnHere) {
-                returnHere.hidden = !detached;
-                if (detached) setButtonDisabled(returnHere, Boolean(state.transfer
-                    && !isClosedWindow(popup) && state.transfer.phase !== "complete"));
-            }
-            if (returnHere && !detached) returnHere.textContent = "Return chat here";
-            if (showWindow) showWindow.hidden = !detached;
-            if (!detached) {
-                const heading = placeholder?.querySelector("h3");
-                const message = placeholder?.querySelector("p");
-                if (heading) heading.textContent = "Chat is open in another window";
-                if (message) message.textContent = "Use the separate Chat window to continue this conversation.";
-            }
-            notifyDetachedChange();
-            if (detached) focusPlaceholderAction();
-        }
-
-        function restoreSourceAfterFailure() {
-            setDetachedUi(false);
-            if (mainLayout) chatUi.restoreLayout?.(mainLayout);
-            mainLayout = null;
-            try { target.focus(); } catch (error) { logger?.debug?.("Main Chat focus failed", error); }
-            const previousFocus = originalFocus;
-            originalFocus = null;
-            focusFirst(target.document.getElementById("chat-input"), previousFocus,
-                target.document.getElementById("btn-chat-new"),
-                target.document.getElementById("btn-open-history"),
-                target.document.getElementById("btn-chat-popout"));
-        }
-
-        function showRecoveryMessage() {
-            const placeholder = target.document.getElementById("chat-window-placeholder");
-            const heading = placeholder?.querySelector("h3");
-            const message = placeholder?.querySelector("p");
-            const showWindow = target.document.getElementById("btn-chat-show-window");
-            const returnHere = target.document.getElementById("btn-chat-return-here");
-            const focusWasOnShowWindow = target.document.activeElement === showWindow;
-            if (heading) heading.textContent = "The Chat window was closed";
-            if (message) message.textContent = "Recover the saved Chat workspace here when you are ready.";
-            if (showWindow) showWindow.hidden = true;
-            if (returnHere) {
-                returnHere.hidden = false;
-                setButtonDisabled(returnHere, false);
-                returnHere.textContent = "Recover chat here";
-            }
-            if (focusWasOnShowWindow) focusPlaceholderAction();
-        }
-
-        function showObserverRecovery(reason) {
-            const placeholder = target.document.getElementById("chat-window-placeholder");
-            const layout = target.document.getElementById("chat-layout");
-            const heading = placeholder?.querySelector("h3");
-            const message = placeholder?.querySelector("p");
-            const showWindow = target.document.getElementById("btn-chat-show-window");
-            const returnHere = target.document.getElementById("btn-chat-return-here");
-            if (layout) layout.hidden = true;
-            if (placeholder) placeholder.hidden = false;
-            if (heading) heading.textContent = "Chat is open in another window";
-            if (message) message.textContent = reason || "Close the other Chat window, then recover the saved workspace here.";
-            if (showWindow) showWindow.hidden = true;
-            if (returnHere) {
-                returnHere.hidden = false;
-                setButtonDisabled(returnHere, false);
-                returnHere.textContent = "Recover chat here";
-            }
-        }
-
-        function updateControls(state) {
-            const transferState = coordinator.getTransferState();
-            const popout = target.document.getElementById("btn-chat-popout");
-            if (popout) {
-                const popoutDisabled = detached || !coordinator.isOwner() || state.popoutAvailable !== true
-                    || transferState.allowed !== true;
-                setButtonDisabled(popout, popoutDisabled);
-                popout.title = popoutDisabled
-                    ? (transferState.reason || state.reason || DEFAULT_POPOUT_TITLE)
-                    : DEFAULT_POPOUT_TITLE;
-            }
-            const returnHere = target.document.getElementById("btn-chat-return-here");
-            if (returnHere && detached) {
-                setButtonDisabled(returnHere, Boolean(state.transfer && !isClosedWindow(popup)
-                    && state.transfer.phase !== "complete"));
-            }
-            if (state.status === "detached" && !detached) setDetachedUi(true);
-            if (state.ownership && detached) {
-                detached = false;
-                setDetachedUi(false);
-                if (mainLayout) chatUi.restoreLayout?.(mainLayout);
-                focusMainChatFallback();
-                mainLayout = null;
-            }
-        }
-
-        async function openPopout() {
-            if (isClosedWindow(popup) === false) {
-                try { popup.focus(); } catch (error) { logger?.debug?.("Chat popout focus failed", error); }
-                return true;
-            }
-            setHostStatus("");
-            const state = coordinator.getState();
-            const transferState = coordinator.getTransferState();
-            if (!state.popoutAvailable || !coordinator.isOwner() || transferState.allowed !== true) {
-                setDetachedUi(false, transferState.reason || state.reason || "Chat popout is unavailable in this browser.");
-                return false;
-            }
-            if (!storage || typeof storage.setItem !== "function") {
-                setDetachedUi(false, "Chat popout requires available same-origin storage.");
-                setHostStatus("Chat remains available in this window. Separate Chat windows require available same-origin storage.");
-                return false;
-            }
-            const key = `llama-gui:chat-window-probe:${randomId("probe", { window: target })}`;
-            const value = randomId("partition", { window: target });
-            try { storage.setItem(key, value); } catch (error) {
-                logger?.debug?.("Chat popout storage probe failed", error);
-                setDetachedUi(false, "Chat popout requires available same-origin storage.");
-                setHostStatus("Chat remains available in this window. Separate Chat windows require available same-origin storage.");
-                return false;
-            }
-            popupProof = { key, value };
-            let opened = null;
-            try {
-                originalFocus = target.document.activeElement;
-                const url = new URL(target.location.href);
-                url.searchParams.set(DETACHED_QUERY_PARAM, "1");
-                url.searchParams.delete("preset");
-                opened = target.open(url.href, POPUP_NAME, POPUP_FEATURES);
-            } catch (error) {
-                logger?.debug?.("Chat popout could not be opened", error);
-            }
-            if (!opened) {
-                try { storage.removeItem(key); } catch (error) { logger?.debug?.("Chat popout probe cleanup failed", error); }
-                popupProof = null;
-                originalFocus = null;
-                setDetachedUi(false, "Allow popups for this page to open Chat in a separate window.");
-                setHostStatus("Chat remains available in this window. Allow popups for this page to open Chat in a separate window.");
-                return false;
-            }
-            popup = opened;
-            mainLayout = chatUi.captureLayout?.() || null;
-            coordinator.attachPeer(popup, { origin: getOrigin({ window: target }), sessionId: coordinator.sessionId });
-            if (!closedCheckTimer && typeof setInterval === "function") {
-                closedCheckTimer = setInterval(() => {
-                    if (!isClosedWindow(popup)) return;
-                    popup = null;
-                    popupProof = null;
-                    if (detached) showRecoveryMessage();
-                    notifyDetachedChange();
-                    if (closedCheckTimer) { clearInterval(closedCheckTimer); closedCheckTimer = null; }
-                }, 500);
-            }
-            const ready = await waitForPeer(coordinator, 10000);
-            if (!ready || isClosedWindow(popup)) {
-                try { storage.removeItem(key); } catch (error) { logger?.debug?.("Chat popout probe cleanup failed", error); }
-                popupProof = null;
-                if (!isClosedWindow(popup)) {
-                    try { popup.close(); } catch (error) { logger?.debug?.("Failed Chat popup close", error); }
-                }
-                popup = null;
-                restoreSourceAfterFailure();
-                setHostStatus("Chat remains available in this window. The separate Chat window could not connect, so the current workspace was preserved.");
-                return false;
-            }
-            try { storage.removeItem(key); } catch (error) { logger?.debug?.("Chat popout probe cleanup failed", error); }
-            if (!await coordinator.beginTransfer({ timeoutMs: 10000 })) {
-                popupProof = null;
-                if (coordinator.getState().status === "recovery-required") {
-                    setDetachedUi(true, "Chat ownership needs explicit recovery from the saved workspace.");
-                    try { target.focus(); } catch (error) { logger?.debug?.("Main Chat focus failed", error); }
-                } else {
-                    try { popup?.close?.(); } catch (error) { logger?.debug?.("Failed Chat popup close", error); }
-                    popup = null;
-                    restoreSourceAfterFailure();
-                    setHostStatus("Chat remains available in this window. The separate Chat window could not take ownership, so the current workspace was preserved.");
-                }
-                return false;
-            }
-            popupProof = null;
-            setHostStatus("");
-            setDetachedUi(true);
-            if (isClosedWindow(popup)) showRecoveryMessage();
-            return true;
-        }
-
-        async function requestReturn() {
-            if (!detached) return false;
-            if (isClosedWindow(popup)) return recoverMain();
-            return coordinator.requestReturn();
-        }
-
-        async function recoverMain() {
-            const locked = await coordinator.acquireOwnership({ ifAvailable: true, activate: false });
-            if (!locked) return false;
-            const record = coordinator.readRecovery();
-            if (!record.ok && record.reason !== "empty" && record.reason !== "invalidated") {
-                if (record.reason !== "invalid-recovery") {
-                    coordinator.releaseOwnership();
-                    return false;
-                }
-                const recoveredInvalid = await coordinator.recover({ clearInvalid: true });
-                if (recoveredInvalid) {
-                    setDetachedUi(false);
-                    mainLayout = null;
-                    focusMainChatFallback();
-                } else {
-                    coordinator.releaseOwnership();
-                }
-                return recoveredInvalid;
-            }
-            let recovered = false;
-            if (record.ok && record.record && !record.record.invalidated) {
-                recovered = await coordinator.recover();
-                if (!recovered) {
-                    coordinator.releaseOwnership();
-                    return false;
-                }
-            } else {
-                if (!coordinator.resetWorkspace()) {
-                    coordinator.releaseOwnership();
-                    return false;
-                }
-                recovered = await coordinator.acquireOwnership({ ifAvailable: true, activate: true });
-            }
-            if (recovered) {
-                setDetachedUi(false);
-                mainLayout = null;
-                focusMainChatFallback();
-            }
-            return recovered;
-        }
-
-        function getBootstrapInfo(candidate) {
-            if (candidate !== popup || isClosedWindow(popup) || !popupProof) return null;
-            return {
-                instanceId: coordinator.instanceId,
-                sessionId: coordinator.sessionId,
-                origin: getOrigin({ window: target }),
-                protocolVersion: PROTOCOL_VERSION,
-                proofKey: popupProof.key,
-                proofValue: popupProof.value,
-            };
-        }
-
-        chatUi.configureWorkspace({
-            checkpoint: snapshot => coordinator.checkpoint(snapshot),
-            invalidate: () => coordinator.invalidateRecovery().ok,
-            onChange: () => updateControls(coordinator.getState()),
-            detachedView: false,
-        });
-        coordinator.subscribe(updateControls);
-        target.document.getElementById("btn-chat-popout")?.addEventListener("click", () => { void openPopout(); });
-        target.document.getElementById("btn-chat-show-window")?.addEventListener("click", () => {
-            if (!isClosedWindow(popup)) popup.focus();
-        });
-        target.document.getElementById("btn-chat-return-here")?.addEventListener("click", () => {
-            void (detached ? requestReturn() : recoverMain());
-        });
-        if (typeof target.addEventListener === "function") {
-            target.addEventListener("pagehide", () => {
-                // A reload mid-handshake would otherwise orphan the storage
-                // probe key, so remove the pending proof before invalidating.
-                try {
-                    if (popupProof) storage.removeItem(popupProof.key);
-                } catch (error) { logger?.debug?.("Chat popout probe cleanup failed", error); }
-                popupProof = null;
-                notifyHostChange({ type: "session-invalidated" });
-            });
-            target.addEventListener("pageshow", event => {
-                // A cached document's host adapter was revoked on pagehide.
-                // Rebuild the host session and reacquire ownership normally.
-                if (event.persisted) target.location.reload();
-            });
-        }
-        // Chat starts with legacy single-window ownership. Revoke it before
-        // init() so the first render cannot write over a recovery snapshot;
-        // coordinator.initialize() then acquires and restores under the lock.
-        chatUi.setOwnership?.(false);
-        const ready = (async () => {
-            try {
-                if (typeof options.initializeChat === "function") await options.initializeChat();
-                const outcome = await coordinator.initialize({ acquire: true, recover: true });
-                if (!outcome.ok && (outcome.reason === "lock-busy" || outcome.reason === "recovery-failed"
-                    || outcome.reason === "invalid-recovery" || outcome.reason === "recovery-reset-failed")) {
-                    showObserverRecovery(outcome.reason === "lock-busy"
-                        ? "Chat is active in another window. Close it, then recover the saved workspace here."
-                        : "The saved Chat workspace needs recovery. Choose Recover chat here to start with a safe workspace.");
-                }
-                return outcome;
-            } catch (error) {
-                try { coordinator.dispose(); } catch (disposeError) {
-                    logger?.debug?.("Unable to dispose Chat coordinator after startup failure", disposeError);
-                }
-                setHostStatus("Chat is unavailable in this window. Reload this page to restore Chat.");
-                try {
-                    showDetachedError(target, "Chat unavailable", "Chat could not start in this window. Other GUI sections remain available; reload this page to restore Chat and runtime controls.");
-                } catch (showError) {
-                    logger?.warn?.("Unable to show Chat startup failure", showError);
-                }
-                logger?.warn?.("Chat host startup failed", error);
-                return result(false, "chat-init-failed");
-            }
-        })();
-        api._hostView = { coordinator, hostAdapter, openPopout, requestReturn, getBootstrapInfo, notifyHostChange,
-            isDetached: () => detached, ready,
-            get popup() { return popup; } };
-        return ready;
-    }
-
-    async function startDetachedView(options = {}) {
-        const target = options.window || (typeof window !== "undefined" ? window : null);
-        if (!target) return result(false, "chat-host-unavailable");
-        const holder = { coordinator: null, hostCheckTimer: null };
-        try {
-            return await startDetachedViewImpl(options, holder);
-        } catch (error) {
-            if (holder.hostCheckTimer !== null) clearInterval(holder.hostCheckTimer);
-            holder.coordinator?.dispose?.();
-            debug({ window: target }, "Detached Chat bootstrap failed", error);
-            try {
-                showDetachedError(target, "Chat window unavailable", "This Chat window could not connect to the main GUI. Close it, then choose Recover chat here in the main Chat window.");
-            } catch (showError) {
-                debug({ window: target }, "Unable to show detached Chat error", showError);
-            }
-            return result(false, "detached-init-failed");
-        }
-    }
-
-    async function startDetachedViewImpl(options = {}, holder = {}) {
-        const target = options.window || (typeof window !== "undefined" ? window : null);
-        if (!target) return result(false, "chat-host-unavailable");
-        await whenDomReady(target);
-        target.document?.body?.classList.add("chat-window-detached");
-        const opener = target && target.opener;
-        const chatUi = options.chatUi || target?.LlamaGui?.chatUi;
-        let openerChatWindow = null;
-        try { openerChatWindow = opener?.LlamaGui?.chatWindow || null; } catch (error) {
-            debug({ window: target }, "Unable to inspect Chat opener", error);
-        }
-        if (!opener || !chatUi || !openerChatWindow) {
-            showDetachedError(target, "Chat window unavailable", "This page was opened without a verified connection to the main GUI. Open the full GUI in this browser, then choose Chat and Pop out; a GUI opened in another browser cannot share this Chat.", { fullGuiLink: true });
-            return result(false, "chat-host-unavailable");
-        }
-        let descriptor;
-        try { descriptor = openerChatWindow.getBootstrapInfo(target); } catch (error) { descriptor = null; }
-        if (!descriptor || descriptor.origin !== getOrigin({ window: target }) || !descriptor.proofKey || !descriptor.proofValue) {
-            showDetachedError(target, "Chat window unavailable", "Close this window, then choose Recover chat here in the main Chat window.");
-            return result(false, "chat-host-unavailable");
-        }
-        const storage = getStorage({ window: target });
-        let proofMatches = false;
-        try { proofMatches = storage && storage.getItem(descriptor.proofKey) === descriptor.proofValue; } catch (error) { proofMatches = false; }
-        if (!proofMatches) {
-            showDetachedError(target, "Chat window unavailable", "This window could not verify the main Chat storage partition.");
-            return result(false, "storage-partition-mismatch");
-        }
-        const remoteState = { settings: {}, runtime: null, status: null, inference: null };
-        let remoteAdapter = null;
-        let hostLossStarted = false;
-        let hostCheckTimer = null;
-        let coordinator;
-        function markHostUnavailable(error) {
-            remoteAdapter = null;
-            if (hostCheckTimer !== null) {
-                clearInterval(hostCheckTimer);
-                hostCheckTimer = null;
-            }
-            const hostStatus = target.document?.getElementById("chat-window-host-status");
-            if (hostStatus) {
-                hostStatus.textContent = "The main Chat window is unavailable. This workspace is paused; return to the main window to recover. Interrupted data may be checkpointed; messages will not be resent automatically.";
-                hostStatus.hidden = false;
-            }
-            if (error && target.console?.debug) target.console.debug("Chat host bridge became unavailable", error);
-            if (coordinator && !hostLossStarted) {
-                hostLossStarted = true;
-                void coordinator.invalidateSession().catch(reason => target.console?.debug?.("Chat host quarantine failed", reason));
-            } else {
-                chatUi.setHostAvailable?.(false);
-            }
-        }
-        function hasCurrentHostSession() {
-            if (!remoteAdapter || hostLossStarted) return false;
-            try {
-                const current = !isClosedWindow(opener)
-                    ? opener?.LlamaGui?.chatWindow?.getSessionInfo?.(target) : null;
-                if (current && current.origin === descriptor.origin && current.sessionId === descriptor.sessionId
-                    && current.valid === true && remoteAdapter.isSessionValid?.() === true) return true;
-            } catch (error) {
-                target.console?.debug?.("Unable to verify current Chat host session", error);
-            }
-            markHostUnavailable(new Error("The main Chat session changed or closed."));
-            return false;
-        }
-        const remoteHost = {
-            getSettings: () => hasCurrentHostSession() ? remoteAdapter.getSettings() : Object.assign({}, remoteState.settings),
-            setSettings: patch => {
-                if (!hasCurrentHostSession()) throw new Error("Chat host is not connected.");
-                return remoteAdapter.setSettings(patch);
-            },
-            getAuthorizationHeaders: (...args) => {
-                if (!hasCurrentHostSession()) throw new Error("Chat host is not connected.");
-                return remoteAdapter.getAuthorizationHeaders(...args);
-            },
-            resetInferenceBaseline: () => hasCurrentHostSession() ? remoteAdapter.resetInferenceBaseline() : undefined,
-            navigate: (...args) => hasCurrentHostSession() ? remoteAdapter.navigate(...args) : false,
-            isSessionValid: hasCurrentHostSession,
-        };
-        const bridge = createFlagCoreBridge(remoteHost, { window: target, onUnavailable: markHostUnavailable });
-        let returnPromise = null;
-        async function transferToMain() {
-            if (returnPromise) return returnPromise;
-            if (!coordinator || !coordinator.isOwner() || !hasCurrentHostSession()) return false;
-            const allowed = coordinator.getTransferState();
-            if (!allowed || allowed.allowed !== true) return false;
-            returnPromise = coordinator.beginTransfer({ destinationId: descriptor.instanceId, timeoutMs: 10000 })
-                .then(returned => {
-                    if (returned) {
-                        try { opener.focus(); } catch (error) { target.console?.debug?.("Main Chat focus failed", error); }
-                        try { target.close(); } catch (error) { target.console?.debug?.("Chat window close failed", error); }
-                    }
-                    return returned;
-                })
-                .finally(() => { returnPromise = null; });
-            return returnPromise;
-        }
-        function updateReturnControl() {
-            const button = target.document?.getElementById("btn-chat-return");
-            if (!button || !coordinator) return;
-            const allowed = coordinator.getTransferState();
-            const disabled = !coordinator.isOwner() || allowed.allowed !== true;
-            button.disabled = disabled;
-            button.setAttribute("aria-disabled", String(disabled));
-            button.title = allowed.reason || "Return Chat to the main window";
-        }
-        const updateRemote = payload => {
-            const next = payload && typeof payload === "object" ? payload : {};
-            if (next.settings && typeof next.settings === "object") {
-                remoteState.settings = Object.assign({}, next.settings);
-                bridge.refresh(remoteState.settings);
-            }
-            if (Object.prototype.hasOwnProperty.call(next, "runtime")) remoteState.runtime = next.runtime;
-            if (Object.prototype.hasOwnProperty.call(next, "status")) remoteState.status = next.status;
-            if (Object.prototype.hasOwnProperty.call(next, "inference")) remoteState.inference = next.inference;
-            if (next.change?.type === "session-invalidated") {
-                markHostUnavailable();
-            }
-            chatUi.refreshSidebarUI?.();
-            chatUi.updateStatusBadge?.();
-            if (remoteState.inference && options.monitorUi) {
-                options.monitorUi.renderInferenceSnapshot?.(remoteState.inference);
-                options.monitorUi.renderStatsBarFromSnapshot?.(remoteState.inference, target.document);
-            }
-        };
-        coordinator = createCoordinator({
-            window: target,
-            chatUi,
-            sessionId: descriptor.sessionId,
-            onHostUpdate: updateRemote,
-            onReturnRequest: () => { void transferToMain(); },
-        });
-        holder.coordinator = coordinator;
-        if (typeof target.addEventListener === "function") {
-            target.addEventListener("pagehide", () => {
-                const transferState = coordinator.getState().transfer;
-                if (transferState?.phase === "complete") return;
-                void coordinator.invalidateSession().catch(error => target.console?.debug?.("Chat popup quarantine failed", error));
-            });
-        }
-        chatUi.configure({
-            flagCore: bridge,
-            confirmAction: options.confirmAction || target.confirmAction || ((title, message) => target.confirm(title, message)),
-            getLatestStatus: () => remoteState.status,
-            getLifecycleSnapshot: () => remoteState.runtime,
-            snapshotStatsBaseline: remoteHost.resetInferenceBaseline,
-            switchTab: remoteHost.navigate,
-            getApiAuthorizationHeaders: remoteHost.getAuthorizationHeaders,
-        });
-        chatUi.configureWorkspace({
-            checkpoint: snapshot => coordinator.checkpoint(snapshot),
-            invalidate: () => coordinator.invalidateRecovery().ok,
-            onChange: updateReturnControl,
-            detachedView: true,
-        });
-        const initialized = await coordinator.initialize({ acquire: false, recover: false });
-        if (!coordinator.getState().lockAvailable) {
-            await whenDomReady(target);
-            showDetachedError(target, "Chat popout unavailable", "This browser does not support the exclusive lock required for a shared Chat.");
-            return result(false, "locks-unavailable");
-        }
-        // The popup starts as an observer like the host path above: revoke
-        // ownership before init() so nothing is mutable before the verified
-        // handoff grants it. Host availability is granted after the handshake.
-        chatUi.setOwnership?.(false);
-        chatUi.setHostAvailable?.(false);
-        await whenDomReady(target);
-        options.themeUi?.init?.();
-        chatUi.init();
-        target.document.getElementById("btn-chat-return")?.removeAttribute("hidden");
-        target.document.getElementById("btn-chat-return")?.addEventListener("click", () => { void transferToMain(); });
-        coordinator.attachPeer(opener, {
-            origin: descriptor.origin,
-            peerId: descriptor.instanceId,
-            sessionId: descriptor.sessionId,
-        });
-        coordinator.beginHandshake({ proof: { key: descriptor.proofKey, value: descriptor.proofValue } });
-        const verified = await waitForPeer(coordinator, 10000);
-        if (!verified) {
-            chatUi.setHostAvailable?.(false);
-            showDetachedError(target, "Chat host unavailable", "Return to the main window and open Chat again.");
-            return result(false, "host-handshake-failed");
-        }
-        remoteAdapter = openerChatWindow.getPeerHostAdapter(target);
-        if (!remoteAdapter) {
-            chatUi.setHostAvailable?.(false);
-            showDetachedError(target, "Chat host unavailable", "Close this window, then choose Recover chat here in the main Chat window.");
-            return result(false, "host-adapter-unavailable");
-        }
-        updateRemote({
-            settings: remoteAdapter.getSettings(), runtime: remoteAdapter.getRuntime(),
-            status: remoteAdapter.getStatus(), inference: remoteAdapter.getInference(),
-        });
-        chatUi.setHostAvailable?.(true);
-        const hostStatus = target.document.getElementById("chat-window-host-status");
-        if (hostStatus) hostStatus.hidden = true;
-        updateReturnControl();
-        options.monitorUi?.renderInferenceSnapshot?.(remoteState.inference);
-        options.monitorUi?.renderStatsBarFromSnapshot?.(remoteState.inference, target.document);
-        hostCheckTimer = setInterval(hasCurrentHostSession, 500);
-        holder.hostCheckTimer = hostCheckTimer;
-        return Object.assign(result(true, "detached-ready"), { coordinator, hostAdapter: remoteAdapter, initialized });
-    }
-
-    const api = {
-        PROTOCOL, PROTOCOL_VERSION, SNAPSHOT_VERSION, SNAPSHOT_KIND, RECOVERY_VERSION,
-        DEFAULT_LOCK_NAME, DEFAULT_RECOVERY_KEY, CHAT_READONLY_SETTING_FIELDS,
-        createHostAdapter, createCoordinator,
-        isDetachedView,
-        startHostView,
-        startDetachedView,
-        notifyHostChange(change) {
-            return api._hostView ? api._hostView.notifyHostChange(change) : false;
-        },
-        getPeerHostAdapter(candidate) {
-            return api._hostView?.coordinator.getPeerHostAdapter(candidate) || null;
-        },
-        getSessionInfo(candidate) {
-            const view = api._hostView;
-            if (!view || candidate !== view.popup) return null;
-            return Object.assign(view.coordinator.getSessionInfo(), { valid: view.hostAdapter.isSessionValid() });
-        },
-        hasDetachedView: () => Boolean(api._hostView && api._hostView.isDetached?.()
-            && !isClosedWindow(api._hostView.popup)),
-        abortActiveStream: () => api._hostView ? api._hostView.coordinator.abortActiveStream() : Promise.resolve(false),
-        getBootstrapInfo(candidate) {
-            return api._hostView?.getBootstrapInfo(candidate) || null;
-        },
-        configure(options) {
-            if (!api._coordinator) api._coordinator = createCoordinator(options);
-            return api._coordinator.initialize(options);
-        },
-        get coordinator() { return api._coordinator || null; },
-    };
-
-    root.LlamaGui.chatWindow = api;
+    I.coordinator = Object.freeze({ createCoordinator });
 })();
